requirements.txt
Flask==2.0.1
Flask-Mail==0.9.1
Flask-CORS==3.0.10
python-dotenv==0.19.0
gunicorn==20.1.0 